+++ test/integration/main.test.ts
@@ -1,21 +1,20 @@
 import {
   RunOptions,
-  RunTarget,
-  deleteAllFakedDirs,
+  RunTarget
 } from "github-action-ts-run-api";
-import { mkdirSync, readFileSync, rm, rmSync, writeFileSync } from "node:fs";
+import { mkdirSync, readFileSync, rmSync, writeFileSync } from "node:fs";
 import path from "path";
-import { afterAll, beforeEach, describe, expect, test } from "vitest";
+import { beforeEach, describe, expect, test } from "vitest";
 
 const target = RunTarget.mainJs("action.yml");
 const tmpDir = "./tmp";
-const options = RunOptions.create()
+const options = () => RunOptions.create()
+  .setOutputOptions({parseStderrCommands: true, parseStdoutCommands: true})
   .setInputs({
-    oldVersion: "1.0.1",
     newVersion: "1.1.0",
   })
   .setEnv({
-    TODAY_DATE: "2023-12-02"
+    TODAY_DATE: "2023-12-02",
   })
   .setGithubContext({
     payload: {
@@ -25,22 +24,24 @@ const options = RunOptions.create()
   })
   .setTempDir(tmpDir)
   .setWorkspaceDir(`${tmpDir}/workspace`)
-  .setFakeFsOptions({ rmFakedTempDirAfterRun: false, rmFakedWorkspaceDirAfterRun: false });
+  .setFakeFsOptions({
+    rmFakedTempDirAfterRun: false,
+    rmFakedWorkspaceDirAfterRun: false,
+  });
 
 describe("main", () => {
+  const tmpChangelogFilePath = path.join("tmp/workspace", "CHANGELOG.md");
   beforeEach(async () => {
     rmSync(tmpDir, { force: true, recursive: true });
-    mkdirSync(`${tmpDir}/workspace`, {recursive:true});
+    mkdirSync(`${tmpDir}/workspace`, { recursive: true });
   });
 
   test("should update changelog", async () => {
-    const tmpChangelogFilePath = path.join('tmp/workspace', "CHANGELOG.md");
-
     writeFileSync(
       tmpChangelogFilePath,
       readFileSync("./test/resource/CHANGELOG.md")
     );
-    const result = await target.run(options);
+    const result = await target.run(options());
 
     expect(result.durationMs).lessThanOrEqual(1000);
     expect(result.stdout).toEqual("");
@@ -49,7 +50,25 @@ describe("main", () => {
     expect(result.runnerWarnings.length).toBe(0);
     expect(result.exitCode).toBe(0);
 
-    const current = readFileSync(tmpChangelogFilePath).toString('utf-8');
+    const current = readFileSync(tmpChangelogFilePath).toString("utf-8");
+    expect(current).toMatchSnapshot();
+  });
+
+  test("when first release changelog, should update", async () => {
+    writeFileSync(
+      tmpChangelogFilePath,
+      readFileSync("./test/resource/changelog_first_release.md")
+    );
+    const result = await target.run(options());
+
+    expect(result.durationMs).lessThanOrEqual(1000);
+    expect(result.stdout).toEqual("");
+    expect(result.stderr).toEqual("");
+    expect(result.commands.errors).toEqual([]);
+    expect(result.runnerWarnings.length).toBe(0);
+    expect(result.exitCode).toBe(0);
+
+    const current = readFileSync(tmpChangelogFilePath).toString("utf-8");
     expect(current).toMatchSnapshot();
   });
 });